Address review comments
Fixed server to properly ignore CodecFactory and its inner classes.

Moved HeapBytesCompressor and HeapBytesDecompressor back up to the CodecFactory class.

Moved DirectCodecPool into DirectCodecFactory to hide access from public API.

Removed reflection based implementation of the DirectDecompressor interface for snappy.

Refactored constructors and creation methods to consistently accept CodecName as their parameter.
